refactor: move lock/unlock wallet api calls from child component to parent
Wallet component should not make api calls; moved one level higher to Wallets component
